NF: Add package manager to plugin manager dlg
